implemented turso

+++ src/handlers/db/models.rs
@@ -1,8 +1,7 @@
-// TODO
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
-#[repr(u8)]
+#[repr(u32)]
 pub enum SanitizerMode {
     Automatic = 0,
     ManualEmote = 1,
@@ -16,8 +15,8 @@ impl Default for SanitizerMode {
     }
 }
 
-impl From<u8> for SanitizerMode {
-    fn from(value: u8) -> Self {
+impl From<u32> for SanitizerMode {
+    fn from(value: u32) -> Self {
         match value {
             0 => SanitizerMode::Automatic,
             1 => SanitizerMode::ManualEmote,
@@ -29,7 +28,7 @@ impl From<u8> for SanitizerMode {
 }
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
-#[repr(u8)]
+#[repr(u32)]
 pub enum DeletePermission {
     AuthorAndMods = 0,
     Everyone = 1,
@@ -42,8 +41,8 @@ impl Default for DeletePermission {
     }
 }
 
-impl From<u8> for DeletePermission {
-    fn from(value: u8) -> Self {
+impl From<u32> for DeletePermission {
+    fn from(value: u32) -> Self {
         match value {
             0 => DeletePermission::AuthorAndMods,
             1 => DeletePermission::Everyone,
